Development (#40)
* fix: include hashPrefix configuration parameter

AngularJS 1.6.0 change the hashPrefix default to the bang operator.  To
support this as well as any other custom value that may be specified by
the application developer we chose to allow this to be specified in the
configuration object as part of the application initialization process.

* fix: linting errors

* feat: configurable token endpoint callback timeout

This closes #38 and closes #39.